refactor: coupon과 관련된 유틸 함수 2개의 파일로 분리
